# system2/static/swagger.json
{"swagger": "2.0", "basePath": "/", "paths": {"/api/pessoa_fisica/": {"post": {"responses": {"201": {"description": "Pessoa f\u00edsica cadastrada com sucesso."}}, "summary": "Cadastra uma nova pessoa fisica", "operationId": "post_individual_collection", "tags": ["api/pessoa_fisica"]}, "get": {"responses": {"200": {"description": "Success"}}, "summary": "Retorna a lista de pessoas", "operationId": "get_individual_collection", "parameters": [{"in": "query", "description": "Filtra registros pela data de nascimento", "name": "data_nascimento", "type": "string"}, {"in": "query", "description": "Filtra registros pela data de cadastro", "name": "data_cadastro", "type": "string"}, {"in": "query", "description": "Filtra registro pelo CPF", "name": "cpf", "type": "string"}], "tags": ["api/pessoa_fisica"]}}, "/api/pessoa_fisica/{id_pessoa}/bens": {"parameters": [{"name": "id_pessoa", "in": "path", "required": true, "type": "integer"}], "post": {"responses": {"404": {"description": "Pessoa n\u00e3o encontrada."}}, "summary": "Cadastra um novo bem", "operationId": "post_individual", "tags": ["api/pessoa_fisica"]}, "get": {"responses": {"404": {"description": "Pessoa n\u00e3o encontrada."}}, "operationId": "get_individual", "tags": ["api/pessoa_fisica"]}}, "/api/pessoa_fisica/{id_pessoa}/bens/{id}": {"parameters": [{"name": "id_pessoa", "in": "path", "required": true, "type": "integer"}, {"name": "id", "in": "path", "required": true, "type": "integer"}], "delete": {"responses": {"404": {"description": "Bem n\u00e3o encontrad."}}, "operationId": "delete_individual", "tags": ["api/pessoa_fisica"]}, "get": {"responses": {"404": {"description": "Bem n\u00e3o encontrad."}}, "operationId": "get_individual", "tags": ["api/pessoa_fisica"]}}, "/api/pessoa_fisica/{id_pessoa}/endereco": {"parameters": [{"name": "id_pessoa", "in": "path", "required": true, "type": "integer"}], "get": {"responses": {"404": {"description": "Pessoa n\u00e3o encontrada."}}, "operationId": "get_individual", "tags": ["api/pessoa_fisica"]}, "put": {"responses": {"404": {"description": "Pessoa n\u00e3o encontrada."}, "204": {"description": "Endere\u00e7o atualizado com sucesso."}}, "summary": "Atualiza endere\u00e7o", "operationId": "put_individual", "tags": ["api/pessoa_fisica"]}}, "/api/pessoa_fisica/{id_pessoa}/fontes_de_renda": {"parameters": [{"name": "id_pessoa", "in": "path", "required": true, "type": "integer"}], "post": {"responses": {"404": {"description": "Pessoa n\u00e3o encontrada."}, "204": {"description": "fonte de rende adicionada com sucesso."}}, "summary": "Cadastra uma nova fonte de renda", "operationId": "post_individual", "tags": ["api/pessoa_fisica"]}, "get": {"responses": {"404": {"description": "Pessoa n\u00e3o encontrada."}}, "operationId": "get_individual", "tags": ["api/pessoa_fisica"]}}, "/api/pessoa_fisica/{id_pessoa}/fontes_de_renda/{id}": {"parameters": [{"name": "id_pessoa", "in": "path", "required": true, "type": "integer"}, {"name": "id", "in": "path", "required": true, "type": "integer"}], "delete": {"responses": {"404": {"description": "Bem n\u00e3o encontrad."}}, "operationId": "delete_individual", "tags": ["api/pessoa_fisica"]}, "get": {"responses": {"404": {"description": "Bem n\u00e3o encontrad."}}, "operationId": "get_individual", "tags": ["api/pessoa_fisica"]}}, "/api/pessoa_fisica/{id}": {"parameters": [{"name": "id", "in": "path", "required": true, "type": "integer"}], "delete": {"responses": {"404": {"description": "Pessoa n\u00e3o encontrada."}, "204": {"description": "Pessoa f\u00edsica deletada com sucesso."}}, "summary": "Deleta pessoa f\u00edsica", "operationId": "delete_individual", "tags": ["api/pessoa_fisica"]}, "get": {"responses": {"404": {"description": "Pessoa n\u00e3o encontrada."}}, "operationId": "get_individual", "tags": ["api/pessoa_fisica"]}, "put": {"responses": {"404": {"description": "Pessoa n\u00e3o encontrada."}, "204": {"description": "Pessoa f\u00edsica atualizada com sucesso."}}, "summary": "Atualiza pessoa f\u00edsica", "operationId": "put_individual", "tags": ["api/pessoa_fisica"]}}}, "info": {"title": "API", "version": "1.0"}, "produces": ["application/json"], "consumes": ["application/json"], "tags": [{"name": "api/pessoa_fisica", "description": "API de pessoa f\u00edsica"}], "responses": {"ParseError": {"description": "When a mask can't be parsed"}, "MaskError": {"description": "When any error occurs on mask"}}}